app/services/automation_service.py
# -*- coding: utf-8 -*-
"""
自動化服務 - 處理截圖、壓縮等自動化操作
"""

import os
import time
import logging
import pyautogui
from PIL import Image
from typing import Optional, Dict, Any, Tuple
from datetime import datetime
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.chrome.options import Options
from app.models import AutomationLog
from app.services.email_service import EmailService
from config.settings import Config

logger = logging.getLogger(__name__)

class AutomationService:
    """自動化服務類別"""

    def __init__(self):
        """初始化自動化服務"""
        self.screenshot_dir = Config.SCREENSHOT_DIR
        self.chrome_driver_path = Config.CHROME_DRIVER_PATH
        self.email_service = EmailService()

        # 確保截圖目錄存在
        os.makedirs(self.screenshot_dir, exist_ok=True)

        # 設定 pyautogui
        pyautogui.FAILSAFE = True
        pyautogui.PAUSE = 1

    def take_screenshot(self, filename: Optional[str] = None) -> Optional[str]:
        """
        截取螢幕截圖

        Args:
            filename: 檔案名稱，如果不提供則自動生成

        Returns:
            截圖檔案路徑，失敗則返回None
        """
        try:
            if not filename:
                timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
                filename = f"screenshot_{timestamp}.png"

            filepath = os.path.join(self.screenshot_dir, filename)

            # 截取螢幕截圖
            screenshot = pyautogui.screenshot()
            screenshot.save(filepath)

            logger.info(f"截圖成功: {filepath}")
            return filepath

        except Exception as e:
            logger.error(f"截圖時發生錯誤: {str(e)}")
            return None

    def automate_top_up_process(self, user_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        自動化儲值流程

        Args:
            user_data: 用戶資料和儲值資訊

        Returns:
            執行結果
        """
        start_time = time.time()
        execution_log = []

        try:
            execution_log.append("開始自動化儲值流程")

            # 步驟1: 截圖記錄開始狀態
            start_screenshot = self.take_screenshot(f"topup_start_{int(time.time())}.png")
            execution_log.append("截圖記錄開始狀態")

            # 步驟2: 模擬自動化流程
            execution_log.append(f"處理用戶儲值: 金額 {user_data.get('amount', 0)}")

            # 模擬處理時間
            time.sleep(2)

            execution_time = time.time() - start_time

            return {
                'success': True,
                'message': '自動化儲值流程完成',
                'execution_time': execution_time,
                'log': execution_log,
                'screenshot_path': start_screenshot
            }

        except Exception as e:
            execution_time = time.time() - start_time
            error_message = f"自動化儲值流程失敗: {str(e)}"

            logger.error(error_message)
            execution_log.append(error_message)

            return {
                'success': False,
                'message': error_message,
                'execution_time': execution_time,
                'log': execution_log
            }
